+++ inveroryManagerApp/Models/DatabaseConnection.cs
@@ -2,12 +2,12 @@ using System.Data.SqlClient;
 using System.Globalization;
 
 namespace inveroryManagerApp.Models;
-public class DatabaseConnection
+public partial class DatabaseConnection
 {
     private SqlConnection ConnectToDatabase()
     {
-        string connectionString = @"Data Source=(local);Initial Catalog=invetory_manager;Integrated Security=true";
-        // string connectionString = @"Data Source=.\DEVLON_LOCAL;Initial Catalog=invetory_manager;Trusted_Connection=True;Integrated Security=True";
+        // string connectionString = @"Data Source=(local);Initial Catalog=invetory_manager;Integrated Security=true";
+        string connectionString = @"Data Source=.\DEVLON_LOCAL;Initial Catalog=invetory_manager;Trusted_Connection=True;Integrated Security=True";
 
         SqlConnection conn = new SqlConnection(connectionString);
         conn.Open();
@@ -75,75 +75,6 @@ public class DatabaseConnection
         return staffModels;
     }
 
-    public List<ItemModel> GetListOfItems()
-    {
-        SqlConnection conn = ConnectToDatabase();
-
-        String sql = "SELECT c.company_id, c.company_name, c.company_description, c.created_date, " +
-            "s.supplier_id, s.supplier_name, s.supplier_contact_number, s.created_date, " +
-            "ic.item_category_id, ic.item_category_name, ic.item_category_description, " +
-            "i.item_id, i.item_name, i.item_description, i.acquired_date, i.cost_price, i.sell_price, i.quantity, i.expiry_date " +
-            "FROM [invetory_manager].[dbo].[Item] as i " +
-            "INNER JOIN [invetory_manager].[dbo].[Company] as c " +
-            "ON i.fk_company_id = c.company_id " +
-            "INNER JOIN [invetory_manager].[dbo].[Supplier] as s " +
-            "ON i.fk_supplier_id = s.supplier_id " +
-            "INNER JOIN [invetory_manager].[dbo].[ItemCategory] as ic " +
-            "ON i.fk_item_category_id = ic.item_category_id";
-        SqlCommand command = new SqlCommand(sql, conn);
-        SqlDataReader dataReader = command.ExecuteReader();
-        List<ItemModel> itemModels = new List<ItemModel>();
-        CultureInfo culture = new CultureInfo("en-US");
-
-        while (dataReader.Read())
-        {
-            CompanyModel company = new CompanyModel() { 
-                CompanyId = (int)dataReader.GetValue(0),
-                CompanyName = dataReader.GetValue(1).ToString(), 
-                CompanyDescription = dataReader.GetValue(2).ToString(),
-                CreatedDate = Convert.ToDateTime(dataReader.GetValue(3).ToString(), culture)
-            };
-            SupplierModel supplier = new SupplierModel()
-            {
-                SupplierId = (int)dataReader.GetValue(4),
-                SupplierName = dataReader.GetValue(5).ToString(),
-                SupplierContactNumber = dataReader.GetValue(6).ToString(),
-                DateAdded = Convert.ToDateTime(dataReader.GetValue(7).ToString(), culture)
-            };
-            ItemCategoryModel itemCategory = new ItemCategoryModel()
-            {
-                ItemCategoryId = (int)dataReader.GetValue(8),
-                ItemCategoryName = dataReader.GetValue(9).ToString(),
-                ItemCategoryDescription = dataReader.GetValue(10).ToString()
-            };
-
-            ItemModel itemModel = new ItemModel()
-            {
-                ItemId = (int)dataReader.GetValue(11),
-                Company = company,
-                Supplier = supplier,
-                ItemCategory = itemCategory,
-                ItemName = dataReader.GetValue(12).ToString(),
-                ItemDescription = dataReader.GetValue(13).ToString(),
-                AcquiredDate = Convert.ToDateTime(dataReader.GetValue(14).ToString(), culture),
-                CostPrice = (decimal)dataReader.GetValue(15),
-                SellPrice = (decimal)dataReader.GetValue(16),
-                Quantity = (int)dataReader.GetValue(17),
-            };
-
-            if (dataReader.GetValue(18).ToString() != "")
-                itemModel.ExpiryDate = Convert.ToDateTime(dataReader.GetValue(18).ToString(), culture);
-            else
-                itemModel.ExpiryDate = null;
-
-            itemModels.Add(itemModel);
-        }
-
-        CloseConnectionToDatabase(conn);
-
-        return itemModels;
-    }
-
     public List<CompanyModel> GetListOfCompany() {
         SqlConnection conn = ConnectToDatabase();
 
@@ -200,34 +131,6 @@ public class DatabaseConnection
 
         return supplierModels;
     }
-
-    public List<ItemCategoryModel> GetItemCategoryModels()
-    {
-        SqlConnection conn = ConnectToDatabase();
-
-        String sql = "SELECT ic.item_category_id, ic.item_category_name, ic.item_category_description " +
-            "FROM [invetory_manager].[dbo].[ItemCategory] as ic";
-        SqlCommand command = new SqlCommand(sql, conn);
-        SqlDataReader dataReader = command.ExecuteReader();
-        List<ItemCategoryModel> itemCategoryModels = new List<ItemCategoryModel>();
-        CultureInfo culture = new CultureInfo("en-US");
-
-        while (dataReader.Read())
-        {
-            ItemCategoryModel itemCategory = new ItemCategoryModel()
-            {
-                ItemCategoryId = (int)dataReader.GetValue(0),
-                ItemCategoryName = dataReader.GetValue(1).ToString(),
-                ItemCategoryDescription = dataReader.GetValue(2).ToString()
-            };
-
-            itemCategoryModels.Add(itemCategory);
-        }
-
-        CloseConnectionToDatabase(conn);
-
-        return itemCategoryModels;
-    }
     
     private CompanyModel GetCompanyByCompanyId(int company_id)
     {
@@ -277,101 +180,6 @@ public class DatabaseConnection
 
         CloseConnectionToDatabase(conn);
         return supplier;
-    }
-
-    private ItemCategoryModel GetItemCategoryByItemCategoryId(int item_category_id)
-    {
-        SqlConnection conn = ConnectToDatabase();
-
-        String sql = "SELECT ic.item_category_id, ic.item_category_name, ic.item_category_description " +
-            "FROM [invetory_manager].[dbo].[ItemCategory] as ic " +
-            "WHERE ic.item_category_id = " + item_category_id;
-        SqlCommand command = new SqlCommand(sql, conn);
-        SqlDataReader dataReader = command.ExecuteReader();
-        CultureInfo culture = new CultureInfo("en-US");
-
-        ItemCategoryModel itemCategory = new ItemCategoryModel();
-
-        while (dataReader.Read())
-        {
-            itemCategory.ItemCategoryId = (int)dataReader.GetValue(0);
-            itemCategory.ItemCategoryName = dataReader.GetValue(1).ToString();
-            itemCategory.ItemCategoryDescription = dataReader.GetValue(2).ToString();
-        }
-
-        CloseConnectionToDatabase(conn);
-        return itemCategory;
-    }
-
-    public ItemModel GetItemByItemId(int item_id)
-    {
-        SqlConnection conn = ConnectToDatabase();
-
-        String sql = "SELECT c.company_id, c.company_name, c.company_description, c.created_date, " +
-            "s.supplier_id, s.supplier_name, s.supplier_contact_number, s.created_date, " +
-            "ic.item_category_id, ic.item_category_name, ic.item_category_description, " +
-            "i.item_id, i.item_name, i.item_description, i.acquired_date, i.cost_price, i.sell_price, i.quantity, i.expiry_date " +
-            "FROM [invetory_manager].[dbo].[Item] as i " +
-            "INNER JOIN [invetory_manager].[dbo].[Company] as c " +
-            "ON i.fk_company_id = c.company_id " +
-            "INNER JOIN [invetory_manager].[dbo].[Supplier] as s " +
-            "ON i.fk_supplier_id = s.supplier_id " +
-            "INNER JOIN [invetory_manager].[dbo].[ItemCategory] as ic " +
-            "ON i.fk_item_category_id = ic.item_category_id " +
-            "WHERE i.item_id = " + item_id;
-        SqlCommand command = new SqlCommand(sql, conn);
-        SqlDataReader dataReader = command.ExecuteReader();
-        CultureInfo culture = new CultureInfo("en-US");
-
-        ItemModel itemModel = new ItemModel();
-
-        while (dataReader.Read())
-        {
-            CompanyModel company = new CompanyModel()
-            {
-                CompanyId = (int)dataReader.GetValue(0),
-                CompanyName = dataReader.GetValue(1).ToString(),
-                CompanyDescription = dataReader.GetValue(2).ToString(),
-                CreatedDate = Convert.ToDateTime(dataReader.GetValue(3).ToString(), culture)
-            };
-            SupplierModel supplier = new SupplierModel()
-            {
-                SupplierId = (int)dataReader.GetValue(4),
-                SupplierName = dataReader.GetValue(5).ToString(),
-                SupplierContactNumber = dataReader.GetValue(6).ToString(),
-                DateAdded = Convert.ToDateTime(dataReader.GetValue(7).ToString(), culture)
-            };
-            ItemCategoryModel itemCategory = new ItemCategoryModel()
-            {
-                ItemCategoryId = (int)dataReader.GetValue(8),
-                ItemCategoryName = dataReader.GetValue(9).ToString(),
-                ItemCategoryDescription = dataReader.GetValue(10).ToString()
-            };
-
-            itemModel.ItemId = (int)dataReader.GetValue(11);
-            itemModel.Company = company;
-            itemModel.Supplier = supplier;
-            itemModel.ItemCategory = itemCategory;
-            itemModel.ItemName = dataReader.GetValue(12).ToString();
-            itemModel.ItemDescription = dataReader.GetValue(13).ToString();
-            itemModel.AcquiredDate = Convert.ToDateTime(dataReader.GetValue(14).ToString(), culture);
-            itemModel.CostPrice = (decimal)dataReader.GetValue(15);
-            itemModel.SellPrice = (decimal)dataReader.GetValue(16);
-            itemModel.Quantity = (int)dataReader.GetValue(17);
-
-            if (dataReader.GetValue(18).ToString() != "")
-                itemModel.ExpiryDate = Convert.ToDateTime(dataReader.GetValue(18).ToString(), culture);
-            else
-                itemModel.ExpiryDate = null;
-
-            itemModel.listCompany = company.CompanyId;
-            itemModel.listSupplier = supplier.SupplierId;
-            itemModel.listItemCategory = itemCategory.ItemCategoryId;
-        }
-
-        CloseConnectionToDatabase(conn);
-
-        return itemModel;
     }
 
     public CustomerModel GetCustomerByCustomerId(int customer_id)
@@ -428,27 +236,6 @@ public class DatabaseConnection
         CloseConnectionToDatabase(conn);
     }
 
-    public void AddItem(ItemModel item)
-    {
-        if (item.ItemName == null || item.listCompany == 0 || item.listSupplier == 0 || item.listItemCategory == 0)
-            return;
-
-        SqlConnection conn = ConnectToDatabase();
-
-        var exp = (item.ExpiryDate?.ToString("'yyyy-MM-dd'") ?? "NULL");
-
-        String sql = "INSERT INTO [invetory_manager].[dbo].[Item] " +
-            "(fk_company_id, fk_supplier_id, fk_item_category_id, item_name, item_description, " +
-            "acquired_date, cost_price, sell_price, quantity, expiry_date) " +
-            "VALUES (" + item.listCompany + ", " + item.listSupplier + ", " + item.listItemCategory + ", '" + item.ItemName + "', '" + item.ItemDescription + "', " +
-            "'" + item.AcquiredDate.ToString("yyyy-MM-dd") + "', " + item.CostPrice + ", " + item.SellPrice + ", " + item.Quantity + ", " + (item.ExpiryDate?.ToString("'yyyy-MM-dd'") ?? "NULL") + "); " +
-            "SELECT SCOPE_IDENTITY()";
-        SqlCommand command = new SqlCommand(sql, conn);
-        int insertedID = Convert.ToInt32(command.ExecuteScalar());
-
-        CloseConnectionToDatabase(conn);
-    }
-
     public void AddCompany(CompanyModel company)
     {
         if (company.CompanyName == null)
@@ -479,68 +266,6 @@ public class DatabaseConnection
         CloseConnectionToDatabase(conn);
     }
 
-    public void AddItemCategory(ItemCategoryModel itemCategory)
-    {
-        if (itemCategory.ItemCategoryName == null)
-            return;
-
-        SqlConnection conn = ConnectToDatabase();
-
-        String sql = "INSERT INTO [invetory_manager].[dbo].[ItemCategory] (item_category_name, item_category_description)" +
-            "VALUES ('" + itemCategory.ItemCategoryName + "', '" + itemCategory.ItemCategoryDescription + "'); SELECT SCOPE_IDENTITY()";
-        SqlCommand command = new SqlCommand(sql, conn);
-        int insertedID = Convert.ToInt32(command.ExecuteScalar());
-
-        CloseConnectionToDatabase(conn);
-    }
-
-    public List<OrderModel> GetOrders()
-    {
-        SqlConnection conn = ConnectToDatabase();
-        String sql = "SELECT i.item_id, i.item_name, o.order_id, o.quantity, o.discount, o.price_paid " +
-            "FROM [invetory_manager].[dbo].[Orders] as o " +
-            "INNER JOIN [invetory_manager].[dbo].[Item] as i " +
-            "ON o.fk_item_id = i.item_id ";
-        SqlCommand command = new SqlCommand(sql, conn);
-        SqlDataReader dataReader = command.ExecuteReader();
-        List<OrderModel> orderModels = new List<OrderModel>();
-
-        while (dataReader.Read())
-        {
-            ItemModel item = new ItemModel() {
-                ItemId = (int)dataReader.GetValue(0),
-                ItemName = dataReader.GetValue(1).ToString(),
-            };
-
-            OrderModel order = new OrderModel()
-            {
-                OrderId = (int)dataReader.GetValue(2),
-                Item = item,
-                Quantity = (int)dataReader.GetValue(3),
-                Discount = (decimal)dataReader.GetValue(4),
-                PricePaid = (decimal)dataReader.GetValue(5),
-            };
-
-            orderModels.Add(order);
-        }
-        CloseConnectionToDatabase(conn);
-
-        return orderModels;
-    }
-
-    public void AddOrder(OrderModel order)
-    {
-        if (order.listItem == 0 || order.Quantity == 0 || order.Discount == 0 || order.PricePaid == 0)
-            return;
-
-        SqlConnection conn = ConnectToDatabase();
-        String sql = "INSERT INTO [invetory_manager].[dbo].[Orders] ([fk_item_id], [quantity], [discount], [price_paid]) " +
-            "VALUES(" + order.listItem + " ," + order.Quantity + " ," + order.Discount + " ," + order.PricePaid + ")";
-        SqlCommand command = new SqlCommand(sql, conn);
-        int insertedID = Convert.ToInt32(command.ExecuteScalar());
-        CloseConnectionToDatabase(conn);
-    }
-
     public void EditStaff(StaffModel staff)
     {
         SqlConnection conn = ConnectToDatabase();
@@ -561,22 +286,6 @@ public class DatabaseConnection
         String sql = "UPDATE [invetory_manager].[dbo].[Customer] " +
             "SET customer_name = '" + customer.CustomerName + "', customer_contact_number = '" + customer.CustomerContactNumber + "' " +
             "WHERE customer_id = " + customer.CustomerId;
-        SqlCommand command = new SqlCommand(sql, conn);
-        command.ExecuteNonQuery();
-
-        CloseConnectionToDatabase(conn);
-    }
-
-    public void EditItem(ItemModel item)
-    {
-        SqlConnection conn = ConnectToDatabase();
-
-        String sql = "UPDATE [invetory_manager].[dbo].[Item] " +
-            "SET fk_company_id = " + item.listCompany + ", fk_supplier_id = " + item.listSupplier + ", fk_item_category_id = " + item.listItemCategory + ", " +
-            "item_name = '" + item.ItemName + "', item_description = '" + item.ItemDescription + "', acquired_date = '" + item.AcquiredDate.ToString("yyyy-MM-dd") + "', " +
-            "cost_price = " + item.CostPrice + ", sell_price = " + item.SellPrice + ", quantity = " + item.Quantity + ", " +
-            "expiry_date = " + (item.ExpiryDate?.ToString("'yyyy-MM-dd'") ?? "NULL") +  " " + 
-            "WHERE item_id = " + item.ItemId;
         SqlCommand command = new SqlCommand(sql, conn);
         command.ExecuteNonQuery();
