+++ inveroryManagerApp/Views/Orders/ListOrders.cshtml
@@ -13,7 +13,7 @@
                     <th scope="col">Order ID</th>
                     <th scope="col">Item ID</th>
                     <th scope="col">Item Name</th>
-]                    <th scope="col">Quantity</th>
+                    <th scope="col">Quantity</th>
                     <th scope="col">Discount</th>
                     <th scope="col">Price Paid</th>
                 </tr>
